minor bug fixes

+++ bash/start_hdp_dev.sh
@@ -111,7 +111,6 @@ g_CENTOS_VERSION="6.8"
 g_AMBARI_VERSION="2.6.1.5" # TODO: need to update Ambari Version manually
 g_STACK_VERSION="2.6"
 
-
 __PID="$$"
 __LAST_ANSWER=""
 
@@ -635,7 +634,7 @@ function _ambari_blueprint_host_groups() {
 '
         else
             _final_hsot_groups='
-    { "name" : "host_group_1", "components" : ['${_master_comps}','${_standby_comps}','${_slave_comps}${_extra_sec_master_comps}${_extra_sec_slave_comps}'], "configurations" : [ ] }
+    { "name" : "host_group_1", "components" : ['${_master_comps}','${_standby_comps}','${_slave_comps}','${_clients}${_extra_sec_master_comps}${_extra_sec_slave_comps}'], "configurations" : [ ] }
 '
         fi
     elif [ $_how_many = 2 ]; then
@@ -2060,6 +2059,7 @@ function f_ambari_set_repo() {
             _warn "${r_AMBARI_VER}: Ambari 2.6.x and higher need VDF file as URL. Trying to generate URL from hdp_urlinfo.json"
             # always get the latest
             curl -sO http://public-repo-1.hortonworks.com/HDP/hdp_urlinfo.json
+            # http://public-repo-1.hortonworks.com/HDF/hdf_urlinfo.json
             # Get the latest vdf file (confusing, sometimes centos, sometimes redhat
             local _vdf="`python -c 'import json;f=open("hdp_urlinfo.json");j=json.load(f);print j["'${_stack}-${_stack_version}'"]["manifests"]["'${_hdp_version}'"]["'${_os_type}'"]'`" || return $?
             [ -z "${_vdf} " ] && return 1
@@ -2463,7 +2463,9 @@ function f_dnsmasq_banner_reset() {
     rm -rf /tmp/banner_add_hosts
 
     # if no banner file, no point of updating it.
-    scp -q $_dns:/etc/banner_add_hosts /tmp/banner_add_hosts || return $?
+    if [ -s /etc/banner_add_hosts ]; then
+        scp -q $_dns:/etc/banner_add_hosts /tmp/banner_add_hosts || return $?
+    fi
 
     if [ -n "${_docker0}" ]; then
         # If an empty file
@@ -2644,17 +2646,18 @@ function f_ssh_setup() {
     which ssh-keygen &>/dev/null || return $?
 
     if [ ! -e $HOME/.ssh/id_rsa ]; then
-        ssh-keygen -f $HOME/.ssh/id_rsa -q -N ""
+        ssh-keygen -f $HOME/.ssh/id_rsa -q -N "" || return 11
     fi
 
     if [ ! -e $HOME/.ssh/id_rsa.pub ]; then
-        ssh-keygen -y -f $HOME/.ssh/id_rsa > $HOME/.ssh/id_rsa.pub
+        ssh-keygen -y -f $HOME/.ssh/id_rsa > $HOME/.ssh/id_rsa.pub || return 12
     fi
 
     _key="`cat $HOME/.ssh/id_rsa.pub | awk '{print $2}'`"
     grep "$_key" $HOME/.ssh/authorized_keys &>/dev/null
     if [ $? -ne 0 ] ; then
-      cat $HOME/.ssh/id_rsa.pub >> $HOME/.ssh/authorized_keys && chmod 600 $HOME/.ssh/authorized_keys
+        cat $HOME/.ssh/id_rsa.pub >> $HOME/.ssh/authorized_keys && chmod 600 $HOME/.ssh/authorized_keys
+        [ $? -ne 0 ] && return 13
     fi
 
     if [ ! -e $HOME/.ssh/config ]; then
